+++ example/lib/main.dart
@@ -38,10 +38,16 @@ class ExampleScreen extends StatelessWidget {
             //Example usage of PriceText
             PriceText(
               currencyType:
-                  CurrencyType.INR, //for currency type to be displayed
+                  CurrencyType.AUD, //for currency type to be displayed
               amount:
-                  -2000, //for amount to be displayed on screen 0 value to show grey color
+                  12345.67, //for amount to be displayed on screen 0 value to show grey color
               showFlag: true, //to show flag of the country
+              locale: "en_AU", //locale for formatting
+              formatterPattern: '#,##0.00', //custom pattern for formatting
+              amountTextStyle: const TextStyle(
+                fontSize: 22,
+                fontWeight: FontWeight.w500,
+              ), //custom text style for amount
               showCurrencyCode: true, //to show currency code
               contryCodeSpacing: 10, //spacing between flag and currency code
               flagSpacing: 20, //spacing between currency code and amount
@@ -60,13 +66,13 @@ class ExampleScreen extends StatelessWidget {
                   false, //to hide currency symbol if you want to show only amount
               avoidCurrencyFormat: false, //to avoid currency formatting
               customFlagWidget: Image.asset(
-                'assets/india.png',
+                'assets/aud.png',
                 width: 40,
                 height: 40,
                 fit: BoxFit.cover,
               ), //custom flag widget if you want to show custom flag
               currencyCodeText:
-                  'India', //custom currency code text if you want to show custom text
+                  'AUD', //custom currency code text if you want to show custom text.
               currencyCodeTextStyle: const TextStyle(
                 fontSize: 22,
                 fontWeight: FontWeight.w500,